requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
httplib2==0.8
itsdangerous==0.23
keyring==3.5
mekk.rtm==0.6.0
python-dateutil==2.2
simplejson==3.3.3
six==1.5.2
wsgiref==0.1.2
